Refactor Python bindings and add comprehensive benchmarks
- Refactor src/python.rs into modular structure (src/python/*.rs)
- Add Modal cloud benchmark runner for reproducible benchmarks
- Add CLI tool for batch mgzip conversion
- Add Criterion benchmarks for Rust-level profiling
- Fix all Python tests (NiftiImage constructor API)
- Fix examples to use correct API (.to_numpy() instead of .data)
- Clean up documentation (remove AI slop from READMEs)
- Add build.rs for macOS Python linking
- Remove benchmark results from version control (regeneratable)
- Exclude paper/ directory from version control

+++ src/pipeline/lazy.rs
@@ -584,10 +584,10 @@ fn apply_affine(
         .map_err(|e| Error::InvalidDimensions(format!("Shape mismatch: {}", e)))?;
     let mut header = image.header().clone();
     header.ndim = 3;
-    header.dim = [1u16; 7];
-    header.dim[0] = od as u16;
-    header.dim[1] = oh as u16;
-    header.dim[2] = ow as u16;
+    header.dim = [1i64; 7];
+    header.dim[0] = od as i64;
+    header.dim[1] = oh as i64;
+    header.dim[2] = ow as i64;
     header.datatype = DataType::Float32;
     header.scl_slope = 1.0;
     header.scl_inter = 0.0;
